requirements.txt
fastapi
uvicorn
keybert
scikit-learn
playwright
rake-nltk
transformers
faiss-cpu
nltk
streamlit
requests
